Refactor Items to remove redundant getWall, separate wallItems and such.

+++ Game/World/Tile.h
@@ -2,7 +2,7 @@
 #include <cstdint>
 
 struct Tile {
-    uint16_t tileId = 0;
-    uint16_t wallId = 0;
+    uint32_t tileId = 0;
+    uint32_t wallId = 0;
     bool isActive = false;
 };
